webplatform/requirements/base.txt
Django==1.6.5
Unipath==1.0
South==0.8.4
psycopg2==2.5.3
django-braces==1.4.0
djangorestframework==2.3.14